pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-file-reader-mcp"
version = "1.0.0"
description = "MCP server for extracting information from various file types"
authors = ["Document Reader Team <2021122006@yonsei.ac.kr>"]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "mcp>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pymupdf>=1.23.0",
    "pdf2image>=3.1.0",
    "Pillow>=10.0.0",
    "google-generativeai>=0.3.0",
    "python-magic>=0.4.27",
    "chardet>=5.2.0",
    "asyncio-throttle>=1.0.0",
    "pydantic>=2.5.0",
    "structlog>=23.0.0"
]

[project.optional-dependencies]
development = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
universal-file-reader = "document_reader.mcp_server:main"
